+++ src/modules/authentication/authentication-api.module.spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { Issuer } from 'openid-client';
 
 import { ConfigTestModule, DatabaseTestModule, MapperTestModule } from '../../../test/utils/index.js';
-import { AuthenticationController } from './api/frontend.controller.js';
+import { AuthenticationController } from './api/authentication.controller.js';
 import { OIDC_CLIENT } from './services/oidc-client.service.js';
 import { AuthenticationApiModule } from './authentication-api.module.js';
