+++ src/modules/authentication/authentication-api.module.ts
@@ -2,7 +2,7 @@ import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
 import { PassportModule } from '@nestjs/passport';
 import { LoggerModule } from '../../core/logging/logger.module.js';
-import { AuthenticationController } from './api/frontend.controller.js';
+import { AuthenticationController } from './api/authentication.controller.js';
 import { OpenIdConnectStrategy } from './passport/oidc.strategy.js';
 import { SessionSerializer } from './passport/session.serializer.js';
 import { OIDCClientProvider } from './services/oidc-client.service.js';
