+++ src/server/server.module.spec.ts
@@ -1,5 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
+import { Issuer } from 'openid-client';
 import { ServerModule } from './server.module.js';
+import { OIDC_CLIENT } from '../modules/authentication/services/oidc-client.service.js';
 
 describe('ServerModule', () => {
     let module: TestingModule;
@@ -7,7 +9,10 @@ describe('ServerModule', () => {
     beforeAll(async () => {
         module = await Test.createTestingModule({
             imports: [ServerModule],
-        }).compile();
+        })
+            .overrideProvider(OIDC_CLIENT)
+            .useValue(new new Issuer({ issuer: 'oidc' }).Client({ client_id: 'DummyId' }))
+            .compile();
     });
 
     afterAll(async () => {
